hw09: add type for general struct validator
